requirements.txt
google-cloud-aiplatform
google-genai
playwright
beautifulsoup4
fastapi
uvicorn
python-multipart
python-dotenv
streamlit
pandas
pytest
pytest-asyncio